refactor: remove styled-components
