+++ src/app/learn/Top.js
@@ -21,7 +21,7 @@ import { useOpenFilter } from '#/state/application/hooks'
 import { updateOpenFilter } from '#/state/application/reducer'
 import { useAppDispatch } from '#/state/hooks'
 import { ChevronLeftIcon } from '@heroicons/react/20/solid'
-import { baseInputStyles } from '#/styleds'
+import { baseInputStyles } from '#/domain/profile/widgets/blocks'
 import { SearchIcon, FilterIcon } from '@/components/Icons'
 import clsx from 'clsx'
