+++ src/app/profile/page.js
@@ -24,7 +24,7 @@ import { SelectCountry } from './SelectCountry'
 import { MySkill } from './MySkill'
 import SocialSettingsFormView from '#/domain/profile/views/social-settings-form'
 import { Setting } from './Setting'
-import { ProfileTitle, ProfileLabel } from '#/styleds'
+import { ProfileTitle, ProfileLabel } from '#/domain/profile/widgets/blocks'
 import { BASE_INPUT_STYLE } from '@/constants/config'
 import { useUser } from '#/state/application/hooks'
 import { postUserInfo } from '#/services/user'
