+++ src/app/shilling-myself/ShillingMyselfOne.js
@@ -18,7 +18,7 @@
 
 import { useEffect, useMemo, useState } from 'react'
 import { classNames } from '@/utils'
-import { baseInputStyles } from '#/styleds'
+import { baseInputStyles } from '#/domain/profile/widgets/blocks'
 import Switch from '@/components/Switch'
 import { Button } from '@/components/Button'
 import { AspectaIcon, SkillsWalletIcon } from '@/components/Icons'
