+++ src/app/shilling/[id]/Author.js
@@ -21,7 +21,7 @@ import { Button } from '@/components/Button'
 import Avatar from '@/components/Avatar'
 import { Modal } from '@/components/Modal'
 import { useState } from 'react'
-import { baseInputStyles } from '#/styleds'
+import { baseInputStyles } from '#/domain/profile/widgets/blocks'
 import { useAllSkills } from '#/state/application/hooks'
 import { countries } from '#/lib/countries'
 import { useDetailsPermission } from '#/services/shilling/hooks'
