+++ src/shared/components/SelectSkills/index.js
@@ -19,7 +19,7 @@ import { useMemo, useState, Fragment, useEffect } from 'react'
 
 import { Transition } from '@headlessui/react'
 import { classNames, arrRemove } from '@/utils'
-import { baseInputStyles } from '#/styleds'
+import { baseInputStyles } from '#/domain/profile/widgets/blocks'
 import { CheckIcon, XMarkIcon } from '@heroicons/react/20/solid'
 import { useConfig } from '#/state/application/hooks'
